FILTER-4: Add more tests and refactor for Datafilter UI
